Remove dependency on glib. Refs #2.
This package currently provides one operation that makes it depend on Glib.
Although many hails applications need Gtk anyway, not all do, and it becomes
harder to support other platforms (iOS, Android) if we have such dependency.

This commit eliminates that operation, making it a comment, and removes
the dependency from the cabal file.